The matrix is now of type double

+++ A3_2020274/src/DiagonalMatrix.java
@@ -1,12 +1,12 @@
 public class DiagonalMatrix extends TriangularMatrix{
     char ID;
     int column, row;
-    int[][] m;
+    double[][] m;
     String type;
 
     DiagonalMatrix(){
         this.column = this.row = 3;
-        this.m = new int[column][row];
+        this.m = new double[column][row];
         this.type = "Diagonal Matrix";
         for (int i = 0; i < this.m.length; i++) {
             m[i][i] = 1 + i;
